pyproject.toml
[tool.poetry]
name = "presentation api"
version = "0.1.0"
description = ""
authors = ["mary-crkn <mzheng@crkn.ca>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.30.3"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
aiopg = "^1.4.0"
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
bottle = "^0.12.25"
iiif-prezi = "^0.3.0"
jsonschema = "^4.23.0"
jsonpath-rw = "^1.4.0"
python-swiftclient = "^4.6.0"
aiohttp = "^3.10.3"
fastapi-azure-auth = "^5.0.0"
pydantic-settings = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
